pyproject.toml
[project]
name = "portfolio-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.1",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
    "django-import-export"
]
